Inventory System 2.0
Way better, the changes are in Slack, but I'll copy them here as well:
Okay, so all of the inventory functionality is now encapsulated in the prefab Shop Icon.  If you put Shop Icon in any scene, all of the inventory stuff will be there.

The ShopMenu script is no longer being used, all of it's functions are now also done by the ItemManager script.  Remove all references to the ShopMenu script,  It's screwed up EventBehavior, up it should be an easy fix to instead refer to ItemManager.

I tried to add extensive comments to explain what is going on and simplify the code.  Also, I fixed the bug where you could scroll the store or closet menu past where there are no more items.